requirements.txt
# 상품 가공 프로그램 필수 패키지 목록
# 설치 방법: pip install -r requirements.txt

# ============================================
# 데이터 처리
# ============================================
# 중요: numpy를 먼저 설치해야 pandas가 정상 작동합니다!
numpy>=1.24.0,<2.0.0  # 수치 연산 (pandas와 호환되는 버전)
pandas>=2.0.0,<3.0.0  # 데이터 처리 (numpy 1.24+ 필요)
openpyxl>=3.0.0  # Excel 파일 읽기/쓰기

# ============================================
# 이미지 처리
# ============================================
Pillow>=9.0.0  # PIL (이미지 처리)
opencv-python>=4.5.0  # OpenCV (이미지 처리)

# ============================================
# 배경 제거 (Remove BG)
# ============================================
rembg>=2.0.0  # 배경 제거 라이브러리
carvekit>=4.0.0  # 배경 제거 (고급)

# ============================================
# 딥러닝 (GPU 가속용, 선택사항)
# ============================================
# torch>=2.0.0  # PyTorch (GPU 사용 시만 필요)
# torchvision>=0.15.0  # (GPU 사용 시만 필요)
# 주의: GPU 사용 시 CUDA 버전에 맞는 torch 설치 필요
# CPU만 사용: pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu

# ============================================
# API 및 네트워크
# ============================================
openai>=1.0.0  # OpenAI API (Stage 1, 2, 3)
requests>=2.28.0  # HTTP 요청
websocket-client>=1.4.0  # WebSocket (ComfyUI)

# ============================================
# 클라우드 스토리지
# ============================================
boto3>=1.26.0  # AWS S3 / Cloudflare R2

# ============================================
# GUI (Python 기본 포함, 별도 설치 불필요)
# ============================================
# tkinter  # Python 기본 포함

# ============================================
# 기타 (Python 기본 포함)
# ============================================
# pathlib  # Python 3.4+ 기본 포함
# json  # Python 기본 포함
# os, sys, subprocess  # Python 기본 포함

# ============================================
# 개발/빌드 도구 (배포 시 선택사항)
# ============================================
# pyinstaller>=5.13.0  # 실행 파일 생성용

